test: address post-review findings from code quality review
- Move ONNX builder helpers out of conftest.py into tests/_helpers.py so
  test files import from a real module rather than from pytest's plugin file
- Update conftest.py to re-export builders from _helpers
- Fix test_predict.py, test_app.py, test_wandb_registry_pull.py to import
  from _helpers instead of conftest
- Remove duplicated _build_det_onnx from test_cli.py; replace call sites
  with the det_onnx_path fixture
- Fix ckpt_path fixture to reuse lmodule.net so the checkpoint architecture
  always tracks the lmodule fixture rather than diverging silently
- Elevate train_loader_partial / eval_loader_partial to session scope to
  align with the other static fixtures they compose with
- Fix test_readers.py to import via radiologist.utils public API instead of
  radiologist.utils.readers internal submodule
